+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.3.1
 
+* `Shamrock`适配椰奶引用撤回、踢人、禁言、修改群名称、设置管理、取消管理、适配meme表情包
+
+# 1.3.1
+
 * `Shamrock`适配拍一拍方法、适配icqq禁言、全体禁言接口，修改转发
 
 # 1.3.0
