+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lain-plugin",
-  "version": "1.3.1",
+  "version": "1.3.2",
   "author": "Lain.",
   "github": "https://github.com/Zyy955/Lain-plugin",
   "gitee": "https://gitee.com/Zyy955/Lain-plugin",
